fs: add ns_inode_init_owner()
Add a ns_inode_init_owner() helper which can be used to initialize the inode
i_uid and i_gid fields with the caller's fs{g,u}id after shifting them it by
the provided user namespace. The inode_init_owner() helper can simply be
implemented on top of it.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>