stat: add ns_generic_fillattr()
Add a ns_generic_fillattr() helper which can be used to fill in the basic
attribures of a file. In contrast to generic_fillattr() it takes an
additional user namespace argument. The kernel internal struct kstat is
extended to contain to new fields uid_translated and gid_translated.
Callers for ns_generic_fillattr() will receive a struct kstat with i_uid
and i_gid already translated into the passed in user namespace. Arguably
this is how it should've always been done.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>