namei: prepare for idmapped mounts in may_o_create()
If an idmapped mount is encountered may_o_create() will shift inodes according
to the mount's user namespaces before checking for permissions. It will make
sure that the callers fs{g,u}id have a mapping in the user namespace of the
superblock. The latter is authorative when translating into raw device ids.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>