fs: add callback for idmap mount aware filesystems
When the kernel is configured with CONFIG_IDMAP_MOUNTS a few additional inode
callbacks will be provided. A filesystem that is aware of idmapped mounts will
receive the user namespace the mount has been marked with as an additional
argument. This can be used for additional permission checking and also to
enable filesystems to translate between uids and gids if they need to. All
helpers are already there.

If this turns out that idmapped mounts is worthwhile we are happy to also just
extend the existing inode callbacks tree-wide. We need to go through each
filesystem slowly anyway to verify and prepare them to be ready for idmapped
mounts.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>