namei: prepare for idmapped mounts in may_create()
If an idmapped mount is encountered may_create() will check for permissions
after shifting by the user namespace the mount has been marked with and will
make sure that the callers fs{g,u}id have a mapping in the user namespace of
the superblock after the shifting. The latter is authorative when translating
into raw device ids.

Signed-off-by: Christian Brauner <christian.brauner@ubuntu.com>