xattr: add a user namespace argument to xattr_permission()
Change xattr_permission() to reason about permission to reason about xattr
permissions with respect to a user namespace. For now we just use the
initial user namespace, so this is a no-op, but in the next few patches
we'll teach xattr operations about id mapped mounts, and they'll need to
pass the user namespace to the permissions check.

Signed-off-by: Tycho Andersen <tycho@tycho.pizza>