fs: introduce count_open_fds()
In future patches, we'll need a count of the number of open file
descriptors for misc NOFILE cgroup migration, so introduce a helper to do
this.

Note that there is a very similarly named helper called count_open_files().
I don't really understand the logic there, but the answer it gives is not
the number of fds the task currently has open. I don't really understand
why, or if it makes sense to unify these.

Signed-off-by: Tycho Andersen <tandersen@netflix.com>