misc cgroup: introduce an fd counter
This idea is not new [1], but I'm hoping piggybacking on the "new" misc
controller to avoid having to introduce a completely new controller will
make it more tenable.

We are interested primarily in monitoring the number of open fds a
container has. We run a multi-tenant environment on both raw VMs and inside
containers, and this is a metric that people migrating to the
container-based environment miss and would have saved us a few production
issues where tenants were leaking files and didn't realize it.

As a bonus, we can limit it and get an EMFILE instead of ENOMEM, but that
is not our primary motivation.

This patch introduces cgroup migration primitives to the misc cgroup, which
didn't have them before. I didn't know what to do in the face of kvm
migrations, so I left those as no-ops for now. I also did not do any
abstraction of the migration primitives. We are interested in adding other
rlimit-y things to the misc cgroup if this approach looks reasonable, for
the same reasons described above. I tried writing both dynamic-dispatch and
static-dispatch versions, but they introduced a lot of noise that seemed
unnecessary for this first draft. I'm happy to take suggestions here.

I have written one test for the happy path, but am happy to write more
tests if the approach is reasonable.

[1]: https://lore.kernel.org/all/1404311407-4851-1-git-send-email-merimus@google.com/

Signed-off-by: Tycho Andersen <tandersen@netflix.com>